refactor: move adc reading logic to adc
since adc-reading contained logic dependent on hardware implementation, it was moved to adc, and other files are modified to accomodate this change
